# server/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CommerceOS API Documentation",
    "version": "1.0.0",
    "description": "Swagger OpenApi specification for CommerceOS Modular Monolith Backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/appointments/slots": {
      "get": {
        "tags": [
          "Appointments Booking"
        ],
        "summary": "Retrieve availability slots for scheduling visits",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "duration",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "X-Store-Id",
            "in": "header",
            "required": false,
            "description": "Optional explicit tenant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success response wrapper",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "success": true,
                    "data": {},
                    "message": "Success"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/appointments/book": {
      "post": {
        "tags": [
          "Appointments Booking"
        ],
        "summary": "Book an appointment time slot",
        "parameters": [
          {
            "name": "X-Store-Id",
            "in": "header",
            "required": false,
            "description": "Optional explicit tenant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "customerName": "Jane Doe",
                  "customerEmail": "jane@example.com",
                  "customerPhone": "0987654321",
                  "serviceName": "Bridal Set Fitting",
                  "duration": 60,
                  "date": "2026-07-01",
                  "time": "11:00"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response wrapper",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "success": true,
                    "data": {},
                    "message": "Success"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}